+++ Source/ACE.DatLoader/Entity/AnimationPartChange.cs
@@ -2,7 +2,6 @@ using System.IO;
 
 namespace ACE.DatLoader.Entity
 {
-    // TODO: Refactor to merge with existing AnimationOverride object in ACE.
     public class AnimationPartChange : IUnpackable
     {
         public byte PartIndex { get; private set; }